backend/requirements.txt
fastapi
uvicorn
sqlalchemy
kafka-python
websockets
aiokafka
asyncpg
psycopg2-binary
python-dotenv
